requirements.txt
faster-whisper>=1.1.0
python-dotenv>=1.0.0
anthropic>=0.42.0
openai>=1.60.0
customtkinter>=5.2.0
Pillow>=10.0.0
